fix(mcp): remediate JSON parse, scope, rate limit, and type coercion errors
- Phase 1: Switch 11 non-JSON tools from run() to run_raw() + Content::text(),
  add --json fields to list_releases
- Phases 2-3: Add classify_gh_error() helper that enriches missing_scope and
  rate limit errors with account, command, and reset time context
- Phase 4: Add flexible_u64/flexible_u64_opt serde helpers to accept both
  numeric and string-encoded integers from LLM callers; normalize all limit
  fields from Option<u32> to Option<u64>

Design doc: docs/design/2026-03-27-mcp-error-remediation.md

+++ src/tools/code.rs
@@ -1,5 +1,6 @@
 //! Code and content tool request types
 
+use crate::serde::flexible_u64_opt;
 use rmcp::schemars;
 use schemars::JsonSchema;
 use serde::Deserialize;
@@ -41,7 +42,8 @@ pub struct SearchCodeRequest {
 
     /// Maximum number of results
     #[schemars(description = "Maximum number of results (default: 30)")]
-    pub limit: Option<u32>,
+    #[serde(default, deserialize_with = "flexible_u64_opt")]
+    pub limit: Option<u64>,
 }
 
 /// Request parameters for list_commits tool
@@ -73,7 +75,8 @@ pub struct ListCommitsRequest {
 
     /// Maximum number of commits to return
     #[schemars(description = "Maximum number of commits to return (default: 30)")]
-    pub limit: Option<u32>,
+    #[serde(default, deserialize_with = "flexible_u64_opt")]
+    pub limit: Option<u64>,
 }
 
 #[cfg(test)]
